+++ src/main/java/roomescape/model/reservation/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.model.reservation;
 
 public class Reservation {
