+++ src/main/java/roomescape/model/reservation/ReservationService.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.model.reservation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
